+++ vizchemoton/html_module.py
@@ -986,4 +986,8 @@ def get_energy_ref(graph,path_list):
     start_nodes = [path[0] for path in path_list]
     all_energies = [graph.nodes[nd]["energy"] for nd in start_nodes]
     idx = np.argmin(all_energies)
+<<<<<<< HEAD
+=======
+    print("output",start_nodes[idx],all_energies[idx])
+>>>>>>> 35d8407dbda28332eaec5dc9229ebcc6ebf301a2
     return (start_nodes[idx],all_energies[idx])