+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Release 2.1.0
 * Added defaults to the config.yaml to minimize errors when using slightly different input files.
 * Added the option to export the chemical reaction network graph in Gephi format.
 * Added reaction pathways to the HTML output so that the user can examine the most likely mechanism between a source and a target compound. Reaction pathways are precomputed and injected into the HTML dashboard, that is, no dynamic generation of the pathways is possible once the HTML is generated. Nonetheless, reaction pathways are deduced from the JSON and CSV interoperable files, therefore, the user can regenerate a new HTML dashboard with the new alternative reaction pathways (no connection to the MongoDB is needed at this stage).
-
+* Added custom script in `vizchemoton/resources/search_substructure_in_crn.py` to perform substructure search from the command line interface. Future versions will evaluate the integration of substructure search in the HTML. 
 
 Release 2.0.0
 -------------